src/main/java/nextstep/ladder/domain/common/Point.java
package nextstep.ladder.domain.common;

public class Point {

    private final int index;
    private final Direction direction;

    private Point(final int index, final Direction direction) {
        this.index = index;
        this.direction = direction;
    }

    public static Point of(final int index, final Direction direction) {
        return new Point(index, direction);
    }

    public static Point first() {
        return new Point(0, Direction.first());
    }

    public int move() {
        if (direction.isRight()) {
            return index + 1;
        }

        if (direction.isLeft()) {
            return index - 1;
        }

        return this.index;
    }

    public Point next() {
        return new Point(index + 1, direction.next());
    }

    public Point last() {
        return new Point(index + 1, direction.last());
    }
}
